メソッドを別ファイルに分けた
